move newmark to its own file
